cart

+++ database/migrations/2025_01_12_182233_create_transactions_table.php
@@ -11,22 +11,23 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('transactions', function (Blueprint $table) {
-            $table->id('transaction_id'); // Primary Key
-            $table->unsignedBigInteger('order_id'); // Foreign Key (order)
-            $table->unsignedBigInteger('user_id'); // Add user_id column
-            $table->decimal('total_price', 8, 2); // Total price in RM
-            $table->unsignedBigInteger('delivery_id'); // Foreign Key (delivery)
-            $table->string('status'); // Status of the transaction (e.g., pending, completed, etc.)
-            $table->timestamps();
+        if (!Schema::hasTable('transactions')) {
+            Schema::create('transactions', function (Blueprint $table) {
+                $table->id('transaction_id'); // Primary Key
+                $table->unsignedBigInteger('order_id'); // Foreign Key (order)
+                $table->unsignedBigInteger('user_id'); // Add user_id column
+                $table->decimal('total_price', 8, 2); // Total price in RM
+                $table->unsignedBigInteger('delivery_id'); // Foreign Key (delivery)
+                $table->string('status'); // Status of the transaction (e.g., pending, completed, etc.)
+                $table->timestamps();
 
-            // Foreign Key Constraints
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade'); // Add foreign key constraint
-            $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
-            $table->foreign('delivery_id')->references('id')->on('deliveries')->onDelete('cascade');
-        });
+                // Foreign Key Constraints
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade'); // Add foreign key constraint
+                $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
+                $table->foreign('delivery_id')->references('id')->on('deliveries')->onDelete('cascade');
+            });
+        }
     }
-
     /**
      * Reverse the migrations.
      */
